address warning

+++ src/str.c
@@ -55,7 +55,7 @@ char* str_format_va_list(const char* fmt, va_list args) {
     va_list args_copy;
     va_copy(args_copy, args);
     // Get the length of the formatted string.
-    int len = vsnprintf(NULL, 0, fmt, args_copy);
+    size_t len = (size_t)vsnprintf(NULL, 0, fmt, args_copy);
     va_end(args_copy);
     if (len < 0) return NULL;
     // Allocate a buffer for the formatted string.
